CLOUDP-196977 refactor local deployments
